Remove OpenSearchYamlFactory / OpenSearchYamlParser
