+++ gradle/libs.versions.toml
@@ -11,8 +11,8 @@ jts                     = "1.15.0"
 jackson_annotations     = "2.22"
 jackson                 = "2.22.0"
 jackson_databind        = "2.22.0"
-jackson3                = "3.1.4"
-jackson3_databind       = "3.1.4"
+jackson3                = "3.2.0"
+jackson3_databind       = "3.2.0"
 snakeyaml               = "2.6"
 snakeyaml_engine        = "3.0.1"
 icu4j                   = "77.1"
