--- libs/x-content/licenses/jackson-core-3.1.4.jar.sha1
@@ -1,1 +1,0 @@
-f9fd39204bd3d4922b42f6a54b81f46bd8ed575a
