+++ server/licenses/jackson-core-3.2.0.jar.sha1
@@ -1,0 +1,1 @@
+ebf2fc67340907abfc4891ed74ba4c4419e2b47e
