--- server/licenses/jackson-dataformat-cbor-3.1.4.jar.sha1
@@ -1,1 +1,0 @@
-e9725dcf40565f68ba3a4ae36c67beee1a2698cb
