+++ server/licenses/jackson-dataformat-cbor-3.2.0.jar.sha1
@@ -1,0 +1,1 @@
+d4d0f03c25a989f8d0486e588860b8cf52dd4118
