--- server/licenses/jackson-dataformat-smile-3.1.4.jar.sha1
@@ -1,1 +1,0 @@
-99cd14f6e7dc9b19f48ee7fd1371f2df0126b3fb
