+++ server/licenses/jackson-dataformat-smile-3.2.0.jar.sha1
@@ -1,0 +1,1 @@
+c715b05a8691e6f28deb7ebeb88b6138ef6530bf
